+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Console Command "ICD Uptime" which displays a table with the program install date, current uptime, and time since last system restart.
 
 ## [4.0.1] - 2019-08-15
 ### Changed